requirements.txt
flask
twilio
ccxt
python-dotenv
peewee
schedule
gunicorn
python-dotenv